+++ ReleaseNotes.md
@@ -2,6 +2,7 @@
 
 ## v1.12.0
 * **NEW FEATURE:** Added property `RestClient.CustomUserAgent` to configure HTTP user-agent header to use for application.
+* **ENVIRONMENTAL:** Dll properties including version information is now set correctly.
 
 ## v1.11.0
 * **NEW FEATURE:** Property `RestClient.QcsSessionClient` is now exposed through interface `IRestClient`.